+++ demos/bookstore/app/admin.users.tsx
@@ -1,4 +1,4 @@
-import type { RouteHandlers } from '@remix-run/fetch-router'
+import type { Controller } from '@remix-run/fetch-router'
 import { createRedirectResponse as redirect } from '@remix-run/response/redirect'
 
 import { routes } from './routes.ts'
@@ -206,4 +206,4 @@ export default {
 
     return redirect(routes.admin.users.index.href())
   },
-} satisfies RouteHandlers<typeof routes.admin.users>
+} satisfies Controller<typeof routes.admin.users>