+++ demos/sse/app/messages.ts
@@ -1,8 +1,8 @@
-import type { BuildRouteHandler } from '@remix-run/fetch-router'
+import type { BuildAction } from '@remix-run/fetch-router'
 
 import type { routes } from './routes'
 
-export let messages: BuildRouteHandler<'GET', typeof routes.messages> = (context) => {
+export let messages: BuildAction<'GET', typeof routes.messages> = (context) => {
   let limitParam = context.url.searchParams.get('limit')
   let limit = limitParam ? parseInt(limitParam, 10) : null
   if (!limit || !isFinite(limit)) limit = null